Unify dependency graph generation
